[Cairo] Unify all drawing area BackingStore implementation files since the only implementation now is cairo https://bugs.webkit.org/show_bug.cgi?id=265375
Reviewed by Fujii Hironori.

* Source/WebCore/platform/Cairo.cmake:
* Source/WebCore/platform/SourcesCairo.txt:
* Source/WebCore/platform/graphics/cairo/BackingStoreBackendCairo.h: Removed.
* Source/WebCore/platform/graphics/cairo/BackingStoreBackendCairoImpl.cpp: Removed.
* Source/WebCore/platform/graphics/cairo/BackingStoreBackendCairoImpl.h: Removed.
* Source/WebCore/platform/graphics/cairo/CairoUtilities.h:
* Source/WebKit/PlatformGTK.cmake:
* Source/WebKit/PlatformPlayStation.cmake:
* Source/WebKit/PlatformWin.cmake:
* Source/WebKit/SourcesGTK.txt:
* Source/WebKit/UIProcess/CoordinatedGraphics/DrawingAreaProxyCoordinatedGraphics.cpp:
(WebKit::DrawingAreaProxyCoordinatedGraphics::paint):
(WebKit::DrawingAreaProxyCoordinatedGraphics::incorporateUpdate):
* Source/WebKit/UIProcess/CoordinatedGraphics/DrawingAreaProxyCoordinatedGraphics.h:
* Source/WebKit/UIProcess/cairo/BackingStore.cpp:
(WebKit::createCairoImageSurfaceWithFastMalloc):
(WebKit::BackingStore::BackingStore):
(WebKit::BackingStore::paint):
(WebKit::BackingStore::incorporateUpdate):
(WebKit::BackingStore::scroll):
* Source/WebKit/UIProcess/cairo/BackingStore.h:
* Source/WebKit/UIProcess/cairo/BackingStoreCairo.cpp: Removed.
* Source/WebKit/UIProcess/wc/DrawingAreaProxyWC.cpp:
(WebKit::DrawingAreaProxyWC::paint):
(WebKit::DrawingAreaProxyWC::incorporateUpdate):
* Source/WebKit/UIProcess/wc/DrawingAreaProxyWC.h:

Canonical link: https://commits.webkit.org/271193@main
