+++ Source/WebCore/platform/Cairo.cmake
@@ -7,8 +7,6 @@ list(APPEND WebCore_UNIFIED_SOURCE_LIST_FILES
 )
 
 list(APPEND WebCore_PRIVATE_FRAMEWORK_HEADERS
-    platform/graphics/cairo/BackingStoreBackendCairo.h
-    platform/graphics/cairo/BackingStoreBackendCairoImpl.h
     platform/graphics/cairo/CairoUniquePtr.h
     platform/graphics/cairo/CairoOperations.h
     platform/graphics/cairo/CairoUtilities.h
