+++ Source/WebCore/platform/SourcesCairo.txt
@@ -21,7 +21,6 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
 // THE POSSIBILITY OF SUCH DAMAGE.
 
-platform/graphics/cairo/BackingStoreBackendCairoImpl.cpp
 platform/graphics/cairo/CairoUtilities.cpp @no-unify
 platform/graphics/cairo/CairoOperations.cpp
 platform/graphics/cairo/FloatRectCairo.cpp
