+++ Source/WebKit/PlatformPlayStation.cmake
@@ -82,7 +82,6 @@ list(APPEND WebKit_SOURCES
 
     Shared/unix/AuxiliaryProcessMain.cpp
 
-    UIProcess/BackingStore.cpp
     UIProcess/DefaultUndoController.cpp
     UIProcess/LegacySessionStateCodingNone.cpp
     UIProcess/WebGrammarDetail.cpp
@@ -110,7 +109,7 @@ list(APPEND WebKit_SOURCES
 
     UIProcess/WebsiteData/playstation/WebsiteDataStorePlayStation.cpp
 
-    UIProcess/cairo/BackingStoreCairo.cpp
+    UIProcess/cairo/BackingStore.cpp
 
     UIProcess/libwpe/WebPasteboardProxyLibWPE.cpp
 
@@ -152,6 +151,7 @@ list(APPEND WebKit_PRIVATE_INCLUDE_DIRECTORIES
     "${WEBKIT_DIR}/UIProcess/API/libwpe"
     "${WEBKIT_DIR}/UIProcess/API/playstation"
     "${WEBKIT_DIR}/UIProcess/CoordinatedGraphics"
+    "${WEBKIT_DIR}/UIProcess/cairo"
     "${WEBKIT_DIR}/UIProcess/playstation"
     "${WEBKIT_DIR}/WebProcess/WebCoreSupport/curl"
     "${WEBKIT_DIR}/WebProcess/WebPage/CoordinatedGraphics"
