+++ Source/WebKit/PlatformWin.cmake
@@ -72,7 +72,6 @@ list(APPEND WebKit_SOURCES
 
     UIProcess/Automation/cairo/WebAutomationSessionCairo.cpp
 
-    UIProcess/BackingStore.cpp
     UIProcess/DefaultUndoController.cpp
     UIProcess/LegacySessionStateCodingNone.cpp
     UIProcess/WebGrammarDetail.cpp
@@ -88,7 +87,7 @@ list(APPEND WebKit_SOURCES
     UIProcess/WebsiteData/curl/WebsiteDataStoreCurl.cpp
     UIProcess/WebsiteData/win/WebsiteDataStoreWin.cpp
 
-    UIProcess/cairo/BackingStoreCairo.cpp
+    UIProcess/cairo/BackingStore.cpp
 
     UIProcess/wc/DrawingAreaProxyWC.cpp
 
@@ -156,6 +155,7 @@ list(APPEND WebKit_PRIVATE_INCLUDE_DIRECTORIES
     "${WEBKIT_DIR}/UIProcess/Inspector/socket"
     "${WEBKIT_DIR}/UIProcess/Inspector/win"
     "${WEBKIT_DIR}/UIProcess/Plugins/win"
+    "${WEBKIT_DIR}/UIProcess/cairo"
     "${WEBKIT_DIR}/UIProcess/wc"
     "${WEBKIT_DIR}/UIProcess/win"
     "${WEBKIT_DIR}/WebProcess/GPU/graphics/wc"
