+++ Source/WebKit/UIProcess/wc/DrawingAreaProxyWC.cpp
@@ -45,7 +45,7 @@ DrawingAreaProxyWC::DrawingAreaProxyWC(WebPageProxy& webPageProxy)
 {
 }
 
-void DrawingAreaProxyWC::paint(BackingStore::PlatformGraphicsContext context, const WebCore::IntRect& rect, WebCore::Region& unpaintedRegion)
+void DrawingAreaProxyWC::paint(cairo_t* context, const WebCore::IntRect& rect, WebCore::Region& unpaintedRegion)
 {
     unpaintedRegion = rect;
 
@@ -80,7 +80,7 @@ void DrawingAreaProxyWC::incorporateUpdate(UpdateInfo&& updateInfo)
         return;
 
     if (!m_backingStore)
-        m_backingStore.emplace(updateInfo.viewSize, updateInfo.deviceScaleFactor, m_webPageProxy);
+        m_backingStore.emplace(updateInfo.viewSize, updateInfo.deviceScaleFactor);
 
     WebCore::Region damageRegion;
     if (updateInfo.scrollRect.isEmpty()) {
